B9

+++ Models/Order.cs
@@ -20,7 +20,7 @@ namespace e_commerce_web.Models
         public bool Deleted { get; set; }
         public bool Paid { get; set; }
         public DateTime? PaymentDate { get; set; }
-        public int TotalMoney { get; set; }
+        public decimal TotalMoney { get; set; }
         public int? PaymentId { get; set; }
         public string Note { get; set; }
         public string Address { get; set; }
